change time stamps from float to double

+++ src/file_handling_impl/xdf.cpp
@@ -765,7 +765,7 @@ void Xdf::freeUpTimeStamps()
         //we don't need to keep all the time stamps unless it's a stream with irregular samples
         if (stream.info.nominal_srate != 0 && !stream.time_stamps.empty())
         {
-            std::vector<float> nothing;
+            std::vector<double> nothing;
             //however we still need to keep the first time stamp of each stream to decide at which position the signal should start
             nothing.emplace_back(stream.time_stamps.front());
             stream.time_stamps.swap(nothing);
