+++ src/file_handling_impl/xdf.h
@@ -54,7 +54,7 @@ public:
     {
         //! A 2D vector which stores the time series of a stream. Each row represents a channel.
         std::vector<std::vector<float> > time_series;
-        std::vector<float> time_stamps; /*!< A vector to store time stamps. */
+        std::vector<double> time_stamps; /*!< A vector to store time stamps. */
         std::string streamHeader;   /*!< Raw XDF of stream header chunk. */
         std::string streamFooter;   /*!< Raw XDF of stream footer chunk. */
